requirements.txt
requests
requests-html
html2text
progressbar2